requirements.txt
alembic==1.12.0
bcrypt==4.0.1
blinker==1.6.3
click==8.1.7
dnspython==2.4.2
dominate==2.8.0
email-validator==2.1.0.post1
Flask==3.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
loguru==0.7.2
Mako==1.2.4
MarkupSafe==2.1.3
psycopg2==2.9.9
python-dotenv==1.0.0
SQLAlchemy==2.0.22
typing_extensions==4.8.0
visitor==0.1.3
WTForms==3.1.0